fix: correct boxplot API types for wp compatibility
- Change all real(8) declarations to real(wp) in boxplot function signature
- Fixes type compatibility between main API and figure implementation
- Ensures proper parameter forwarding without type conversions
- Maintains consistency with other API functions using wp types

+++ src/fortplot.f90
@@ -182,13 +182,13 @@ contains
         !! Example:
         !!   ! Simple box plot
         !!   call boxplot(measurement_data, label='Measurements')
-        real(8), dimension(:), intent(in) :: data
-        real(8), intent(in), optional :: position
-        real(8), intent(in), optional :: width
+        real(wp), dimension(:), intent(in) :: data
+        real(wp), intent(in), optional :: position
+        real(wp), intent(in), optional :: width
         character(len=*), intent(in), optional :: label
         logical, intent(in), optional :: show_outliers
         logical, intent(in), optional :: horizontal
-        real(8), intent(in), optional :: color(3)
+        real(wp), intent(in), optional :: color(3)
 
         call fig%boxplot(data, position=position, width=width, label=label, &
                         show_outliers=show_outliers, horizontal=horizontal, color=color)
